ref: move search components to search folder
