+++ src/components/search/VueScreenerSearch.vue
@@ -13,9 +13,9 @@
 <script lang="ts" setup>
 import { VueScreener } from '@/interfaces/screener'
 import { computed, ref } from 'vue'
-import { isValidRegExp } from '../utils/regex.utils'
-import UiInput, { InputUI } from './ui/input/Input.vue'
-import { twMerge } from '../utils/tailwind-merge.utils'
+import { isValidRegExp } from '../../utils/regex.utils'
+import UiInput, { InputUI } from '../ui/input/Input.vue'
+import { twMerge } from '../../utils/tailwind-merge.utils'
 
 export type VueScreenerSearchUI = InputUI
